Add files via upload
V1.01.16 - first GitHub release. Script supports the editing of a XML reference file allowing it to remove a Softpaq from being analyzed by HPIA or to replace it with a previous (superseded) version. In addition. The updated file can then be used with the /ReferenceFile command line option. It will allow the creation of a custom, curated, repository with only selected Softpaqs that HPIA can use to install from - using the /OfflineMode option